Don't sign in if already signed in
The MAS sign in task fails if email and password are provided, but
the current user has already signed in, either manually or
during a previous run of the task.

This change introduces an additional check by using the exit code
of the `mas account` command: the exit code 0 indicates that
we are already signed in, 1 that we are not.